renamed File to Storage + map throws to Result
